Reorganizing bytemaker.bittypes into separate modules and creating an associated subpackage
